backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
redis
passlib[bcrypt]
PyJWT
PyMuPDF
genai
